backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
openpyxl==3.1.5
reportlab==4.4.4
num2words==0.5.14
psycopg2-binary==2.9.9
docxtpl==0.16.7
